requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
greenlet==3.2.3
packaging==25.0
pillow==11.3.0
pygame==2.6.1
SQLAlchemy==2.0.41
typing_extensions==4.14.1
